Dockerfile
This dockerfile will create a docker image which will generate an ssh key and pull down the latest source, including fully functional ionic. I will include instructions on how to use it so we can get new developers up and going in seconds (hopefully!). Developed under Ubuntu but should work on most platforms with a slight tweak to the instructions. Will have a play :)